+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## [1.2.0] (IN PROGRESS)
 
-* Import PropTypes from prop-types, not React. Refs STRIPES-427. 
+* Import PropTypes from prop-types, not React. Refs STRIPES-427.
 * Lint cleanup. Fixes UIPFU-3.
+* Add `dataKey` to connected Users component. Fixes UIPFU-4.
 
 ## [1.1.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.1.0) (2017-09-06)